Commented out unnecessary lines, renamed a few mixin references, and made the @Inject use @WrapMethod instead.
