--- magerymodifications-1.21.1/src/main/java/com/madelyndustries/magerymodifications/Mixins/MixinPlugin.java
@@ -1,5 +1,0 @@
-package com.madelyndustries.magerymodifications.Mixins;
-
-public class MixinPlugin {
-
-}
