+++ magerymodifications-1.21.1/src/main/java/com/madelyndustries/magerymodifications/mixins/PocketDimensionRubberbandingKillerMixin.java
@@ -1,17 +1,17 @@
-package com.madelyndustries.magerymodifications.Mixins;
+package com.madelyndustries.magerymodifications.mixins;
 
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
 import io.redspace.ironsspellbooks.capabilities.magic.PocketDimensionManager;
 import net.minecraft.world.level.Level;
 
 @Mixin(PocketDimensionManager.class)
-public class pocketDimensionRubberbandingKillerMixin {
-    @Inject(method="tick",at=@At("HEAD"), cancellable = true)
-    private static void tick(Level level, CallbackInfo ci){
-        ci.cancel();
+public class PocketDimensionRubberbandingKillerMixin {
+    @WrapMethod(method="tick")
+    private static void tick(Level level, Operation<Void> original){
     }
 }
